ai-assistance/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
chromadb>=0.5.0
sentence-transformers>=3.0.0
openai>=1.40.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiofiles>=23.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
httpx>=0.27.0
